+++ bzt/cli.py
@@ -307,6 +307,12 @@ def main():
 
     executor = CLI(parsed_options)
 
+    stdin = sys.stdin.read()
+    if stdin:
+        with NamedTemporaryFile(prefix="stdin_", suffix=".config", delete=False) as fhd:
+            fhd.write(stdin)
+            parsed_configs.append(fhd.name)
+
     try:
         code = executor.perform(parsed_configs)
     except BaseException as exc_top: