+++ site/dat/docs/Changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.4.3 (next)
-
+ - allow passing config as stdin to CLI like `./generate-config.sh | bzt`
 
 ## 0.4.2
  - bump up Gatling version to 2.1.7